privitize client
